chore: move interfaces to separate file
